requirements.txt
tensorflow
tensorflow-model-optimization
pillow